fix(financeiro): alinha tipos id (increments) entre todas migrations
Mismatch detectado ao rodar migrate em MySQL:
- fin_contas_bancarias.id era bigIncrements
- fin_boleto_remessas.id/titulo_id/conta_bancaria_id eram bigInteger
- mas fin_titulos.id e fin_titulo_baixas.* sao increments (INT)

FK constraints quebravam (Referencing column ... incompatible).

Alinhado tudo pra increments/integer unsigned (consistente com tabelas
Onda 1). Rodou DONE em MySQL local: 7/7 migrations RAN.

+++ Modules/Financeiro/Database/Migrations/2026_04_25_140101_create_fin_boleto_remessas_table.php
@@ -18,10 +18,10 @@ class CreateFinBoletoRemessasTable extends Migration
     public function up(): void
     {
         Schema::create('fin_boleto_remessas', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');
             $table->integer('business_id')->unsigned()->index();
-            $table->bigInteger('titulo_id')->unsigned();
-            $table->bigInteger('conta_bancaria_id')->unsigned();
+            $table->integer('titulo_id')->unsigned();
+            $table->integer('conta_bancaria_id')->unsigned();
 
             $table->string('nosso_numero', 30)
                   ->comment('Sequencial gerado pelo sistema; banco confirma no retorno');
